Moved middleware to pkg/server instead to avoid conflict with reference inside integration folder
