+++ pkg/server/middleware/middleware.go
@@ -6,12 +6,13 @@ import (
 	"net/http"
 	"strings"
 
+	"google.golang.org/grpc/metadata"
+
 	"github.com/go-logr/logr"
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/weaveworks/weave-gitops/pkg/logger"
 	"github.com/weaveworks/weave-gitops/pkg/services/auth"
 	"golang.org/x/oauth2"
-	"google.golang.org/grpc/metadata"
 )
 
 type statusRecorder struct {
